+++ src/main/java/com/imaginea/mongodb/services/CollectionService.java
@@ -55,7 +55,7 @@ public interface CollectionService {
    * @throws CollectionException throw super type of
    *         DuplicateCollectionException,InsertCollectionException
    */
-  public String insertCollection(String dbName, String newCollName, boolean capped, int size,
+  public String insertCollection(String dbName, String newCollName, boolean capped, long size,
       int maxDocs, boolean autoIndexId)
       throws DatabaseException, CollectionException, ValidationException;
 
@@ -76,7 +76,7 @@ public interface CollectionService {
    *         DuplicateCollectionException,InsertCollectionException
    */
   public String updateCollection(String dbName, String selectedCollection, String newCollName,
-      boolean capped, int size, int maxDocs, boolean autoIndexId)
+      boolean capped, long size, int maxDocs, boolean autoIndexId)
       throws DatabaseException, CollectionException, ValidationException;
 
   /**
